refactor: extract the sheet widgets from main_screen.dart
main_screen.dart had grown to 1357 lines: the main screen plus six private
widget classes accumulated across earlier features. The original issue
proposed extracting AlertDialog-based dialogs into lib/widgets/, but PR #31
replaced every dialog with inline sheets before this was picked up -- see
the issue's own comment thread, which retargets it at the private widgets
instead.

Moved to lib/widgets/, renamed to drop the now-cross-file leading
underscore (State classes keep theirs, matching normal Flutter convention):

- SheetInitialFocus -- the Web focus-claiming workaround
- NumericSheet -- the Character Limit sheet
- FontSizeSheet -- field, slider and footer
- FontList / FontTile -- the scrolling font picker with its reveal-on-open
  logic
- ToolbarButton
- AboutBodyText / AboutStudy

Every comment moved with its code unedited, including the Web-specific
behavioral notes the issue's comment thread called out explicitly (focus
claiming, arrow-key traversal, scroll-controller driving) -- these were the
output of real debugging sessions and are exactly the context a future
reader would otherwise have to rediscover. Three comments that referenced
the old underscore-prefixed names by name were updated to match the rename.

main_screen.dart is now 730 lines: MainScreen, _MainScreenState, and the
_SheetInitialFocus row-traversal shortcuts map that only that state class
uses.

Closes #8
